+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 ### Added
  - Advanced sign on back writing
+ - Advanced sign styling menu
 
 ### Fixed
  - Thread safeness of peripherals